excel2notion/domain/requirements.txt
# Core
fastapi==0.115.5
uvicorn==0.34.0
python-dotenv==1.0.1

# Notion & Slack
notion-client==2.2.1
slack-sdk==3.34.0

# Data processing
pandas==2.2.3
openpyxl==3.1.5
numpy<2,>=1.24.0

# HTTP
requests==2.32.3

# AI/ML
google-generativeai==0.8.3
openai==1.55.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0
scikit-learn==1.5.2

# Fix urllib3 version
urllib3<2.0.0

PyPDF2==3.0.1
python-multipart==0.0.20